+++ crates/common/src/consts.rs
@@ -6,4 +6,4 @@ pub const BIN_NAME: &str = "moon";
 
 pub const CONFIG_DIRNAME: &str = ".moon";
 
-pub const PROTO_CLI_VERSION: &str = "0.44.2";
+pub const PROTO_CLI_VERSION: &str = "0.44.4";