+++ crates/toolchain/src/detect/project_toolchain.rs
@@ -1,54 +1,64 @@
 use super::languages::{BUN, DENO, NODE};
 use super::project_language::has_language_files;
 use moon_common::Id;
-use moon_config::LanguageType;
+use moon_config::{LanguageType, PlatformType};
+use std::convert::TryFrom;
 use std::path::Path;
 
+/// Return a list of toolchains based on the provided ID.
+pub fn get_project_toolchains(id: &Id, language: &LanguageType) -> Vec<Id> {
+    let mut toolchains = vec![id.to_owned()];
+
+    // Since JS has multiple runtimes, we should inherit JS/TS also
+    if id == "bun" || id == "deno" || id == "node" {
+        toolchains.push(if matches!(language, LanguageType::TypeScript) {
+            Id::raw("typescript")
+        } else {
+            Id::raw("javascript")
+        });
+    }
+    // Otherwise check if we're a supported platform, if not, inherit system
+    else if PlatformType::try_from(id.as_str()).is_err() {
+        toolchains.push(Id::raw("system"));
+    }
+
+    toolchains
+}
+
+// Detect the correct toolchains based on the project's language
+// and what config files exist in the project root.
 pub fn detect_project_toolchains(
     workspace_root: &Path,
     project_root: &Path,
     language: &LanguageType,
-    enabled_toolchains: &[Id],
 ) -> Vec<Id> {
     let mut toolchains = vec![];
 
-    match language {
-        LanguageType::JavaScript | LanguageType::TypeScript => {
-            let runtimes = vec![
-                (Id::raw("deno"), DENO),
-                (Id::raw("bun"), BUN),
-                (Id::raw("node"), NODE),
-            ];
+    if matches!(
+        language,
+        LanguageType::JavaScript | LanguageType::TypeScript
+    ) {
+        let runtimes = vec![
+            (Id::raw("deno"), DENO),
+            (Id::raw("bun"), BUN),
+            (Id::raw("node"), NODE),
+        ];
 
-            for (id, files) in runtimes {
-                // Detect in project first
-                if has_language_files(project_root, files) {
-                    toolchains.push(id);
-                    break;
-                }
-
-                // Then in workspace
-                if has_language_files(workspace_root, files) {
-                    toolchains.push(id);
-                    break;
-                }
+        for (id, files) in runtimes {
+            // Detect in project first
+            if has_language_files(project_root, files) {
+                toolchains.push(id);
+                break;
             }
 
-            toolchains.extend(language.get_toolchain_ids());
+            // Then in workspace
+            if has_language_files(workspace_root, files) {
+                toolchains.push(id);
+                break;
+            }
         }
-        other => {
-            toolchains.push(Id::raw(other.to_string()));
-        }
-    };
-
-    let mut toolchains = toolchains
-        .into_iter()
-        .filter(|id| enabled_toolchains.contains(id))
-        .collect::<Vec<_>>();
-
-    if toolchains.is_empty() {
-        toolchains.push(Id::raw("system"));
     }
 
+    toolchains.extend(language.get_toolchain_ids());
     toolchains
 }